requirements.txt
torch==1.6.0
torchvision==0.7.0
pytorch-lightning==1.3.1
tensorboard
torchmetrics
textdistance~=4.2.1
opencv-python~=4.5.2.54
scikit-image~=0.18.1
scikit-learn
ctcdecode
streamlit
psycopg2

notebook==6.3.0
matplotlib~=3.4.1

pillow~=8.1.2
numpy~=1.19.2
tqdm~=4.59.0
wandb